# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nocturne API",
    "version": "1.0.0",
    "description": "REST API for discovering and sharing night-view spots in Seoul."
  },
  "servers": [
    {
      "url": "https://nocturne-htdu.onrender.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/users/me/bookmarks": {
      "get": {
        "summary": "List the authenticated user's bookmarked spots",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/me/likes": {
      "get": {
        "summary": "List the spots the authenticated user has liked",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/me/spots": {
      "get": {
        "summary": "List the spots posted by the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/spots": {
      "get": {
        "summary": "List top 10 spots for the card list",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by season or weather tag"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all"
              ]
            },
            "description": "Use `all` to return every spot for map markers"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create a new spot",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "latitude",
                  "longitude"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Spot created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/spots/map": {
      "get": {
        "summary": "List all spots for the map (no limit, includes easter egg spots)",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by season or weather tag"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/spots/{id}/comments": {
      "get": {
        "summary": "List comments for a spot",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid spot id"
          }
        }
      },
      "post": {
        "summary": "Create a comment on a spot",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created"
          },
          "400": {
            "description": "Invalid spot id or comment content"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/spots/{spotId}/comments/{commentId}": {
      "delete": {
        "summary": "Delete a comment owned by the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spotId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted"
          },
          "400": {
            "description": "Invalid spot id or comment id"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Comment not found or not owned by this user"
          }
        }
      }
    },
    "/api/spots/{id}/likes": {
      "post": {
        "summary": "Like a spot",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Like created"
          },
          "400": {
            "description": "Invalid spot id"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Spot already liked by this user"
          }
        }
      },
      "delete": {
        "summary": "Remove a like from a spot",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Like removed"
          },
          "400": {
            "description": "Invalid spot id"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Like not found"
          }
        }
      }
    },
    "/api/spots/{id}/bookmarks": {
      "post": {
        "summary": "Bookmark a spot",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Bookmark created"
          },
          "400": {
            "description": "Invalid spot id"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Spot already bookmarked by this user"
          }
        }
      },
      "delete": {
        "summary": "Remove a bookmark from a spot",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Bookmark removed"
          },
          "400": {
            "description": "Invalid spot id"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Bookmark not found"
          }
        }
      }
    },
    "/api/spots/{id}": {
      "get": {
        "summary": "Get a spot by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Spot not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a spot (owner only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — not the owner"
          },
          "404": {
            "description": "Spot not found"
          }
        }
      },
      "patch": {
        "summary": "Edit a spot (owner only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "seasonTag": {
                    "type": "string"
                  },
                  "weatherTag": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — not the owner"
          },
          "404": {
            "description": "Spot not found"
          }
        }
      }
    },
    "/api/spots/{id}/photos": {
      "post": {
        "summary": "Upload a photo for a spot (owner only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Photo uploaded"
          },
          "400": {
            "description": "Invalid spot id or no file"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — not the owner"
          },
          "404": {
            "description": "Spot not found"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Log in and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a JWT token"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    }
  },
  "tags": []
}